feat: refactor tinymce editor to sharedComponents
