+++ src/editors/containers/ProblemEditor/components/EditProblemView/QuestionWidget/index.test.jsx
@@ -1,6 +1,7 @@
 /* eslint-disable react/prop-types */
 import React from 'react';
 import { shallow } from 'enzyme';
+import { formatMessage } from '../../../../../../testUtils';
 import { actions, selectors } from '../../../../../data/redux';
 import { QuestionWidget, mapStateToProps, mapDispatchToProps } from '.';
 
@@ -11,22 +12,15 @@ jest.mock('../../../../../data/redux', () => ({
     },
   },
   selectors: {
+    app: {
+      lmsEndpointUrl: jest.fn(state => ({ lmsEndpointUrl: state })),
+      studioEndpointUrl: jest.fn(state => ({ studioEndpointUrl: state })),
+    },
     problem: {
       question: jest.fn(state => ({ question: state })),
     },
   },
 }));
-
-// Per https://github.com/tinymce/tinymce-react/issues/91 React unit testing in JSDOM is not supported by tinymce.
-// Consequently, mock the Editor out.
-jest.mock('@tinymce/tinymce-react', () => {
-  const originalModule = jest.requireActual('@tinymce/tinymce-react');
-  return {
-    __esModule: true,
-    ...originalModule,
-    Editor: () => 'TiNYmCE EDitOR',
-  };
-});
 
 jest.mock('../../../hooks', () => ({
   prepareEditorRef: jest.fn(() => ({
@@ -40,6 +34,10 @@ describe('QuestionWidget', () => {
   const props = {
     question: 'This is my question',
     updateQuestion: jest.fn(),
+    lmsEndpointUrl: 'sOmEvaLue.cOm',
+    studioEndpointUrl: 'sOmEoThERvaLue.cOm',
+    // injected
+    intl: { formatMessage },
   };
   describe('render', () => {
     test('snapshot: renders correct default', () => {
@@ -50,6 +48,16 @@ describe('QuestionWidget', () => {
     const testState = { A: 'pple', B: 'anana', C: 'ucumber' };
     test('question from problem.question', () => {
       expect(mapStateToProps(testState).question).toEqual(selectors.problem.question(testState));
+    });
+    test('lmsEndpointUrl from app.lmsEndpointUrl', () => {
+      expect(
+        mapStateToProps(testState).lmsEndpointUrl,
+      ).toEqual(selectors.app.lmsEndpointUrl(testState));
+    });
+    test('studioEndpointUrl from app.studioEndpointUrl', () => {
+      expect(
+        mapStateToProps(testState).studioEndpointUrl,
+      ).toEqual(selectors.app.studioEndpointUrl(testState));
     });
   });
   describe('mapDispatchToProps', () => {
