+++ src/editors/containers/ProblemEditor/index.test.jsx
@@ -26,6 +26,7 @@ jest.mock('../../data/redux', () => ({
   selectors: {
     app: {
       blockValue: jest.fn(state => ({ blockValue: state })),
+      assets: jest.fn(state => ({ assets: state })),
     },
     problem: {
       problemType: jest.fn(state => ({ problemType: state })),
@@ -45,6 +46,8 @@ describe('ProblemEditor', () => {
     blockFinished: false,
     studioViewFinished: false,
     initializeProblemEditor: jest.fn().mockName('args.intializeProblemEditor'),
+    assets: {},
+    assetsFinished: false,
   };
   describe('snapshots', () => {
     test('renders as expected with default behavior', () => {
@@ -56,11 +59,14 @@ describe('ProblemEditor', () => {
     test('studio view not yet loaded, Spinner appears', () => {
       expect(shallow(<ProblemEditor {...props} studioViewFinished />)).toMatchSnapshot();
     });
+    test('assets not yet loaded, Spinner appears', () => {
+      expect(shallow(<ProblemEditor {...props} assetsFinished />)).toMatchSnapshot();
+    });
     test('renders SelectTypeModal', () => {
-      expect(shallow(<ProblemEditor {...props} blockFinished studioViewFinished />)).toMatchSnapshot();
+      expect(shallow(<ProblemEditor {...props} blockFinished studioViewFinished assetsFinished/>)).toMatchSnapshot();
     });
     test('renders EditProblemView', () => {
-      expect(shallow(<ProblemEditor {...props} problemType="multiplechoiceresponse" blockFinished studioViewFinished />)).toMatchSnapshot();
+      expect(shallow(<ProblemEditor {...props} problemType="multiplechoiceresponse" blockFinished studioViewFinished assetsFinished/>)).toMatchSnapshot();
     });
   });
 
@@ -70,6 +76,11 @@ describe('ProblemEditor', () => {
       expect(
         mapStateToProps(testState).blockValue,
       ).toEqual(selectors.app.blockValue(testState));
+    });
+    test('assets from app.assets', () => {
+      expect(
+        mapStateToProps(testState).assets,
+      ).toEqual(selectors.app.assets(testState));
     });
     test('problemType from problem.problemType', () => {
       expect(
@@ -85,6 +96,11 @@ describe('ProblemEditor', () => {
       expect(
         mapStateToProps(testState).studioViewFinished,
       ).toEqual(selectors.requests.isFinished(testState, { requestKey: RequestKeys.fetchStudioView }));
+    });
+    test('assetsFinished from requests.isFinished', () => {
+      expect(
+        mapStateToProps(testState).assetsFinished,
+      ).toEqual(selectors.requests.isFinished(testState, { requestKey: RequestKeys.fetchAssets }));
     });
   });
   describe('mapDispatchToProps', () => {
