+++ src/editors/containers/TextEditor/index.jsx
@@ -1,24 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import { Editor } from '@tinymce/tinymce-react';
-
-import 'tinymce';
-import 'tinymce/themes/silver';
-import 'tinymce/skins/ui/oxide/skin.css';
-import 'tinymce/icons/default';
-import 'tinymce/plugins/link';
-import 'tinymce/plugins/lists';
-import 'tinymce/plugins/table';
-import 'tinymce/plugins/hr';
-import 'tinymce/plugins/codesample';
-import 'tinymce/plugins/emoticons';
-import 'tinymce/plugins/emoticons/js/emojis';
-import 'tinymce/plugins/charmap';
-import 'tinymce/plugins/code';
-import 'tinymce/plugins/autoresize';
-import 'tinymce/plugins/image';
-import 'tinymce/plugins/imagetools';
 
 import {
   Spinner,
@@ -30,11 +12,10 @@ import { actions, selectors } from '../../data/redux';
 import { RequestKeys } from '../../data/constants/requests';
 
 import EditorContainer from '../EditorContainer';
-import ImageUploadModal from './components/ImageUploadModal';
-import SourceCodeModal from './components/SourceCodeModal';
 import RawEditor from '../../sharedComponents/RawEditor';
 import * as hooks from './hooks';
 import messages from './messages';
+import TinyMceWidget from '../../sharedComponents/TinyMceWidget';
 
 export const TextEditor = ({
   onClose,
@@ -52,10 +33,6 @@ export const TextEditor = ({
   intl,
 }) => {
   const { editorRef, refReady, setEditorRef } = hooks.prepareEditorRef();
-  const { isImgOpen, openImgModal, closeImgModal } = hooks.imgModalToggle();
-  const { isSourceCodeOpen, openSourceCodeModal, closeSourceCodeModal } = hooks.sourceCodeModalToggle(editorRef);
-  const images = hooks.filterAssets({ assets });
-  const imageSelection = hooks.selectedImage(null);
 
   if (!refReady) { return null; }
 
@@ -69,20 +46,18 @@ export const TextEditor = ({
       );
     }
     return (
-      <Editor
-        {...hooks.editorConfig({
-          setEditorRef,
-          blockValue,
-          openImgModal,
-          openSourceCodeModal,
-          initializeEditor,
-          lmsEndpointUrl,
-          studioEndpointUrl,
-          isLibrary,
-          images,
-          setSelection: imageSelection.setSelection,
-          clearSelection: imageSelection.clearSelection,
-        })}
+      <TinyMceWidget
+        editorType="text"
+        editorRef={editorRef}
+        textValue={blockValue ? blockValue.data.data : ''}
+        setEditorRef={setEditorRef}
+        minHeight={500}
+        height="100%"
+        initializeEditor={initializeEditor}
+        isLibrary={isLibrary}
+        lmsEndpointUrl={lmsEndpointUrl}
+        studioEndpointUrl={studioEndpointUrl}
+        assets={assets}
       />
     );
   };
@@ -93,21 +68,6 @@ export const TextEditor = ({
       onClose={onClose}
     >
       <div className="editor-body h-75 overflow-auto">
-        {isLibrary ? null : (
-          <ImageUploadModal
-            isOpen={isImgOpen}
-            close={closeImgModal}
-            editorRef={editorRef}
-            images={images}
-            {...imageSelection}
-          />
-        )}
-        <SourceCodeModal
-          isOpen={isSourceCodeOpen}
-          close={closeSourceCodeModal}
-          editorRef={editorRef}
-        />
-
         <Toast show={blockFailed} onClose={hooks.nullMethod}>
           <FormattedMessage {...messages.couldNotLoadTextContext} />
         </Toast>
