+++ src/editors/containers/TextEditor/index.test.jsx
@@ -19,35 +19,18 @@ jest.mock('@tinymce/tinymce-react', () => {
 });
 
 jest.mock('../EditorContainer', () => 'EditorContainer');
+// jest.mock('../../sharedComponents/TinyMceEditor', () => 'TinyMceEditor');
 
-jest.mock('./components/ImageUploadModal', () => 'ImageUploadModal');
-jest.mock('./components/SourceCodeModal', () => 'SourceCodeModal');
 
 jest.mock('./hooks', () => ({
   editorConfig: jest.fn(args => ({ editorConfig: args })),
   getContent: jest.fn(args => ({ getContent: args })),
-  imgModalToggle: jest.fn(() => ({
-    isImgOpen: true,
-    openImgModal: jest.fn().mockName('openModal'),
-    closeImgModal: jest.fn().mockName('closeModal'),
-  })),
-  sourceCodeModalToggle: jest.fn(() => ({
-    isSourceCodeOpen: true,
-    openSourceCodeModal: jest.fn().mockName('openModal'),
-    closeSourceCodeModal: jest.fn().mockName('closeModal'),
-  })),
-  selectedImage: jest.fn(() => ({
-    selection: 'hooks.selectedImage.selection',
-    setSelection: jest.fn().mockName('hooks.selectedImage.setSelection'),
-    clearSelection: jest.fn().mockName('hooks.selectedImage.clearSelection'),
-  })),
   nullMethod: jest.fn().mockName('hooks.nullMethod'),
   prepareEditorRef: jest.fn(() => ({
     editorRef: { current: { value: 'something' } },
     refReady: true,
     setEditorRef: jest.fn().mockName('hooks.prepareEditorRef.setEditorRef'),
   })),
-  filterAssets: jest.fn(() => [{ staTICUrl: '/assets/sOmEaSsET' }]),
 }));
 
 jest.mock('react', () => {
@@ -98,16 +81,6 @@ describe('TextEditor', () => {
     intl: { formatMessage },
   };
   describe('snapshots', () => {
-    imgModalToggle.mockReturnValue({
-      isImgOpen: false,
-      openImgModal: jest.fn().mockName('modal.openModal'),
-      closeImgModal: jest.fn().mockName('modal.closeModal'),
-    });
-    sourceCodeModalToggle.mockReturnValue({
-      isSourceCodeOpen: false,
-      openSourceCodeModal: jest.fn().mockName('modal.openModal'),
-      closeSourceCodeModal: jest.fn().mockName('modal.closeModal'),
-    });
     test('renders as expected with default behavior', () => {
       expect(shallow(<TextEditor {...props} />)).toMatchSnapshot();
     });
@@ -119,9 +92,6 @@ describe('TextEditor', () => {
     });
     test('block failed to load, Toast is shown', () => {
       expect(shallow(<TextEditor {...props} blockFailed />)).toMatchSnapshot();
-    });
-    test('ImageUploadModal is not rendered', () => {
-      expect(shallow(<TextEditor {...props} isLibrary />)).toMatchSnapshot();
     });
   });
 
@@ -136,6 +106,11 @@ describe('TextEditor', () => {
       expect(
         mapStateToProps(testState).lmsEndpointUrl,
       ).toEqual(selectors.app.lmsEndpointUrl(testState));
+    });
+    test('studioEndpointUrl from app.studioEndpointUrl', () => {
+      expect(
+        mapStateToProps(testState).studioEndpointUrl,
+      ).toEqual(selectors.app.studioEndpointUrl(testState));
     });
     test('assets from app.assets', () => {
       expect(
