+++ src/editors/sharedComponents/TinyMceWidget/pluginConfig.js
@@ -3,11 +3,14 @@ import { buttons, plugins } from '../../data/constants/tinyMCE';
 
 const mapToolbars = toolbars => toolbars.map(toolbar => toolbar.join(' ')).join(' | ');
 
-const pluginConfig = (isLibrary) => {
+const pluginConfig = ({ isLibrary, placeholder, editorType }) => {
   const image = isLibrary ? '' : plugins.image;
   const imageTools = isLibrary ? '' : plugins.imagetools;
   const imageUploadButton = isLibrary ? '' : buttons.imageUploadButton;
   const editImageSettings = isLibrary ? '' : buttons.editImageSettings;
+  const codePlugin = editorType === 'problem' ? '' : plugins.code;
+  const codeButton = editorType === 'problem' ? '' : buttons.code;
+  const labelButton = editorType === 'problem' ? buttons.customLabelButton : '';
 
   return (
     StrictDict({
@@ -19,7 +22,7 @@ const pluginConfig = (isLibrary) => {
         plugins.table,
         plugins.hr,
         plugins.charmap,
-        plugins.code,
+        codePlugin,
         plugins.autoresize,
         image,
         imageTools,
@@ -28,6 +31,7 @@ const pluginConfig = (isLibrary) => {
       toolbar: mapToolbars([
         [buttons.undo, buttons.redo],
         [buttons.formatSelect],
+        [labelButton],
         [buttons.bold, buttons.italic, buttons.underline, buttons.foreColor, buttons.backColor],
         [
           buttons.align.left,
@@ -43,7 +47,7 @@ const pluginConfig = (isLibrary) => {
         ],
         [imageUploadButton, buttons.link, buttons.unlink, buttons.blockQuote, buttons.codeBlock],
         [buttons.table, buttons.emoticons, buttons.charmap, buttons.hr],
-        [buttons.removeFormat, buttons.code],
+        [buttons.removeFormat, codeButton],
       ]),
       imageToolbar: mapToolbars([
         // [buttons.rotate.left, buttons.rotate.right],
@@ -54,11 +58,11 @@ const pluginConfig = (isLibrary) => {
         branding: false,
         height: '100%',
         menubar: false,
-        min_height: 500,
         toolbar_sticky: true,
         toolbar_sticky_offset: 76,
         relative_urls: true,
         convert_urls: false,
+        placeholder,
       },
     })
   );
